requirements.txt
openai
python-dotenv
reportlab
requests
duckduckgo-search
